refactor: extract agent notifications from App
